Created map class to quickly create point map from DB

+++ fantasy-colonialism-mapgen/FantasyColonialismMapgen/BiomeGen.cs
@@ -540,7 +540,7 @@ namespace FantasyColonialismMapgen
             public double averageTempSummer;
             public double averageTempWinter;
             public double averageRainfall; // mm/year
-            public int coastalDistance;
+            public double coastalDistance;
             public int length;//EW
             public int width;//NS
         }
